+++ slash/hooks.py
@@ -24,6 +24,8 @@ _define('session_end', doc="Called right before the session ends, regardless of 
 _define('before_session_start', doc="Entry point which is called before session_start, useful for configuring plugins and other global resources")
 _define('after_session_start', doc="Second entry point for session start, useful for plugins relying on other plugins' session_start routine")
 
+_define('configure', doc='Configuration hook that happens during commandline parsing, and before plugins are activated. It is a convenient point to override plugin activation settings')
+
 _define('test_interrupt', doc="Called when a test is interrupted by a KeyboardInterrupt or other similar means")
 _define('test_start', doc="Called right after a test starts")
 _define('test_end', doc="Called right before a test ends, regardless of the reason for termination")